Organise classes into different packages
