+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public abstract class AddCommand implements Command {
     @Override
     public boolean isExit() {
