+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.core.Storage;
+import duke.core.TaskList;
+import duke.core.Ui;
+import duke.exception.DukeException;
 import java.time.format.DateTimeParseException;
 
 public class AddDeadlineCommand extends AddCommand {
@@ -14,7 +20,7 @@ public class AddDeadlineCommand extends AddCommand {
             return tasks.recordDeadline(descriptionDatePair[0].trim(), descriptionDatePair[1].trim());
         } catch (IndexOutOfBoundsException | DateTimeParseException e) {
             throw new DukeException("deadline should be in format: [DESCRIPTION] /by [DATE]!\n" +
-                    "Only accepted [DATE] format is: date/month/year, HHMM (24h time)");
+                    "Only accepted [DATE] format is: date/month/year HHMM (24h time)");
         }
     }
 }
