+++ src/main/java/duke/command/DoneCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.core.Storage;
+import duke.core.TaskList;
+import duke.core.Ui;
+import duke.exception.DukeException;
+
 public class DoneCommand implements Command {
     private String description;
