+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.core.Storage;
+import duke.core.TaskList;
+import duke.core.Ui;
+
 public class ExitCommand implements Command {
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) {
