+++ src/main/java/duke/core/Storage.java
@@ -1,8 +1,11 @@
+package duke.core;
+
+import duke.exception.DukeException;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.Scanner;
 
 public class Storage {
     private String filePath;
