+++ text-ui-test/EXPECTED.TXT
@@ -74,7 +74,7 @@ Nice! I've marked this task as done:
   [E][X] eat lunch (at: Friday,     2pm)
 ------------------------------------------------------------------------
 ------------------------------------------------------------------------
-Task is already marked as done
+duke.task.Task is already marked as done
 ------------------------------------------------------------------------
 ------------------------------------------------------------------------
 --- Start of List ---
